rename PlotWorld to OnePlotWorld and make PlotWorld not abstract
